requirements.txt
asgiref
django-cors-headers
django
djangorestframework
isbnlib
Pillow
pytz
setuptools
sqlparse
psycopg2-binary


